# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthCheck": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "title": "Actorid",
            "type": "string"
          },
          "allowed": {
            "title": "Allowed",
            "type": "boolean"
          },
          "guildId": {
            "title": "Guildid",
            "type": "string"
          }
        },
        "required": [
          "allowed",
          "actorId",
          "guildId"
        ],
        "title": "AuthCheck",
        "type": "object"
      },
...
      "Compatibility": {
        "additionalProperties": false,
        "properties": {
          "apiVersion": {
            "title": "Apiversion",
            "type": "string"
          },
          "coreVersion": {
            "title": "Coreversion",
            "type": "string"
          },
          "schemaVersion": {
            "title": "Schemaversion",
            "type": "integer"
          }
        },
        "required": [
          "coreVersion",
          "apiVersion",
          "schemaVersion"
        ],
        "title": "Compatibility",
        "type": "object"
      },
...
      "Diagnostics": {
        "additionalProperties": false,
        "properties": {
          "botReady": {
            "title": "Botready",
            "type": "boolean"
          },
          "commandSync": {
            "title": "Commandsync",
            "type": "string"
          },
          "driveConfigured": {
            "title": "Driveconfigured",
            "type": "boolean"
          },
          "providerMetrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProviderMetrics"
            },
            "title": "Providermetrics",
            "type": "object"
          },
          "providers": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Providers",
            "type": "object"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "weatherConfigured": {
            "title": "Weatherconfigured",
            "type": "boolean"
          }
        },
        "required": [
          "version",
          "botReady",
          "commandSync",
          "store",
          "providers",
          "providerMetrics",
          "weatherConfigured",
          "driveConfigured"
        ],
        "title": "Diagnostics",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "ProviderMetrics": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "failures": {
            "title": "Failures",
            "type": "integer"
          },
          "lastLatencyMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastlatencyms"
          },
          "successes": {
            "title": "Successes",
            "type": "integer"
          }
        },
        "required": [
          "available",
          "failures",
          "successes",
          "lastLatencyMs"
        ],
        "title": "ProviderMetrics",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ConanGrayBot Core",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/check": {
      "get": {
        "operationId": "auth_check_api_v1_auth_check_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthCheck"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Auth Check"
      }
    },
    "/api/v1/compatibility": {
      "get": {
        "operationId": "compatibility_api_v1_compatibility_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Compatibility"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Compatibility"
      }
    },
...
    "/api/v1/diagnostics": {
      "get": {
        "operationId": "diagnostics_api_v1_diagnostics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagnostics"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Diagnostics"
      }
    },
...
  }
}